Lumpi Lumpi Lumpi

+++ springerdownloader.py
@@ -1,6 +1,10 @@
 import glob
 import urllib.request
+import re
 from bs4 import BeautifulSoup
+import os
+
+special_characters = re.compile(r"[^0-9a-zA-Z_ÄÜÖäüö]")
 
 
 def download_book(url, dir):
@@ -8,13 +12,18 @@ def download_book(url, dir):
     soup = BeautifulSoup(response.read().decode("utf-8"), features="html.parser")
     download_button = soup.find("a", attrs={"data-track-action": "Book download - pdf"})
     title = soup.find("div", attrs={"class": "page-title", "data-test": True}).find("h1").get_text()
-    print("Downloading '" + title + "' from " + url)
-    urllib.request.urlretrieve("https://link.springer.com" + download_button.get("href"),
-                               dir + "/" + title.replace(" ", "_") + ".pdf")
+
+    file_title = re.sub(special_characters, "_", title)
+    file_path = os.path.join(dir, file_title+".pdf")
+    if not os.path.exists(file_path):
+        print("Downloading '" + title + "' from " + url)
+        urllib.request.urlretrieve("https://link.springer.com" + download_button.get("href"),
+                                   file_path)
+    else:
+        print(f"Book '{title}' already exists in '{dir}'.")
 
 
 def main(ddir):
-    import os
     if not os.path.exists(ddir):
         os.makedirs(ddir)
     for file in glob.glob("book_urls/*.txt"):
